Add FUSE_PARALLEL_DIROPS and enable it by default
This is a squished commit of the following two commits with
modification for libfuse 2.

---

Added support for FUSE_PARALLEL_DIROPS

Enabled by default since we haven't released libfuse 3.0 yet :-).

Fixes #112.

---

Pass FUSE_PARALLEL_DIROPS to kernel (#861)

This tells the kernel that parallel lookup/readdir operations are
supported. This is enabled by default but was not passed to the kernel
so you always get the synchronized version.